refactor: files names to meet Google JS guide
